fix: removed polymorphism on commit function
this fixes all the tests and benchmarks that were failing

The previous code also had the implicit assumption that it only worked with polys in eval form. So just made that explicit. We can add the function to commit from coeff form (by taking ifft for eg) later if needed, but opted to keep API simple for now.